requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
google-generativeai>=0.4.0
firebase-admin>=6.4.0
pydantic>=2.6.1
httpx<0.28
